requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-generativeai==0.5.4
duckduckgo-search==6.1.7
